requirements.txt
amqp==5.2.0
asgiref==3.7.2
astroid==2.15.8
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.0
cron_descriptor==1.4.0
distlib==0.3.8
Django==4.2.9
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
drf-yasg==1.21.7
eventlet==0.34.3
filelock==3.13.1
flake8==6.1.0
greenlet==3.0.3
identify==2.5.33
idna==3.6
inflection==0.5.1
kombu==5.3.5
lazy-object-proxy==1.10.0
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
virtualenv==20.25.0
wcwidth==0.2.13
wrapt==1.16.0
